Add search dialog view
